Completed and corrected models, removed easteregg, and included datasheets.
A number of funky hacks allow for 3D model rendering to look good, by
specifying similar shaped packages for lacking models such as LGA-2mm
and 6-pin LED chips. I hacked the rendering even more to represent a
battery clip that's not the one in use at all, which could be misleading.